requirements.txt
anyascii==0.3.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.7.0
beautifulsoup4==4.9.3
boto3==1.24.79
botocore==1.27.79
certifi==2022.6.15
charset-normalizer==2.1.1
Deprecated==1.2.13
dj-database-url==0.4.1
Django==4.0.7
django-cache-url==2.0.0
django-debug-toolbar==3.6.0
django-dotenv==1.4.1
django-filter==21.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-redis==5.2.0
django-storages==1.13.1
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
jmespath==1.0.1
l18n==2021.3
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
toml==0.10.2
urllib3==1.26.12
wagtail==4.0.1
wagtailfontawesome==1.1.4
webencodings==0.5.1
whitenoise==5.0.1
Willow==1.4.1
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
